requirements.txt
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
requests==2.31.0
yfinance==0.2.28
alpha-vantage==2.3.1
ccxt==4.1.64
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0 